Add pixel for page track plus button conversion
